pyproject.toml
[project]
name = "morgenmcp"
version = "0.1.0"
description = "MCP server for Morgen calendar API"
license = { text = "Apache-2.0" }
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=0.4.1",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-core>=2.41.5",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "ruff>=0.9",
    "pre-commit>=4.0",
]

[project.scripts]
morgenmcp = "morgenmcp.server:main"

[tool.uv]
package = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: tests that hit the real Morgen API (deselect with '-m not integration')",
]
addopts = "-m 'not integration'"

[tool.ruff]
target-version = "py314"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]

[tool.ruff.lint.isort]
known-first-party = ["morgenmcp"]
